Move layout context
