+++ app/src/components/GraphWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode, useContext } from "react";
-import { LayoutContext } from "../constants";
 import { Box } from "../slang";
 import styles from "./GraphWrapper.module.css";
+import { LayoutContext } from "./Layout";
 
 export default function GraphWrapper({ children }: { children: ReactNode }) {
   const { showing } = useContext(LayoutContext);
